+++ R/colour_values_hex.R
@@ -23,9 +23,8 @@
 #'   \item{Matrix - }{At least 5 rows, and 3 (or 4) columns representing the red, green and blue (and alpha) values}
 #' }
 #'
-#' The \code{palette} (and \code{alpha} argument when using a vector) requires 5
-#' rows because the colours are interpolated using a cubic b-spline. This method
-#' requires 5 values.
+#' The \code{palette} requires 5 rows because the colours are interpolated using a
+#' cubic b-spline. This method requires 5 values.
 #'
 #' @examples
 #'
@@ -66,7 +65,6 @@
 #'
 #' @export
 colour_values <- function( x, palette = "viridis", na_colour = "#808080FF", alpha = 255, include_alpha = TRUE, ... ) {
-  alpha_check( alpha )
   colour_values_to_hex( x, palette, na_colour, alpha, include_alpha, ... )
 }