+++ tests/testthat/test-colour_values_hex.R
@@ -96,9 +96,10 @@ test_that("alpha values applied", {
 
   expect_true(all(substr(colour_values(1:5, alpha = 0),8,9) == "00"))
 
+  ## No longe valid since v0.2.1 (issue 25)
   #expect_error(colour_values(1:5, alpha = c(100,200)),"alpha must either be a single value, or the same length as x")
-  expect_error(colour_values( 1:5, alpha = c(100,200) ), "A vector of alpha values must contain at least 5 values")
-  expect_error(colour_values( 1:3, alpha = c(1:3)), "A vector of alpha values must contain at least 5 values")
+  #expect_error(colour_values( 1:5, alpha = c(100,200) ), "A vector of alpha values must contain at least 5 values")
+  #expect_error(colour_values( 1:3, alpha = c(1:3)), "A vector of alpha values must contain at least 5 values")
 
   expect_true(all(colour_values(letters) == colour_values(letters, alpha = 255)))
   expect_true(all(substr( colour_values(letters, alpha = 0),8,9) == "00"))